fix store error with reducer
