+++ src/index.js
@@ -1,20 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { createStore } from 'redux';
+
 import './style/index.css';
 import App from './components/App';
-import rootReducer from './reducers/index';
+import store from './reducers/store';
 
-const initialState = [
-  { id: Math.random(), title: 'The Call of the Wild', category: 'Action' },
-  { id: Math.random(), title: 'To Kill a Mockingbird', category: 'Biography' },
-  { id: Math.random(), title: 'And Then There Were None', category: 'History' },
-  { id: Math.random(), title: 'Ninth House', category: 'Horror' },
-  { id: Math.random(), title: 'Clean Code', category: 'Learning' },
-
-];
-const store = createStore(rootReducer, initialState);
 ReactDOM.render(
   <Provider store={store}>
     <React.StrictMode>
